+++ CHANGELOG.md
@@ -11,8 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Use `global`  configuration for injection links
 - Escape  data when check if already exist
-
-## [Unreleased]
+- Fix injection of `values in entity tabs` when injecting an `entity`
 
 ### Added