+++ CHANGELOG.md
@@ -12,10 +12,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Use `global`  configuration for injection links
 - Escape  data when check if already exist
 - Fix injection of `values in entity tabs` when injecting an `entity`
+- Fix `pdffont` field error for users
 
 ### Added
 
 - Add option to `replace` (instead of `append`)  the value of multiline text fields (e.g. `comment`)
+
+### Fixed
+
+- Move `Notepads` search options to the Black List option
 
 ## [2.14.1] - 2024-12-27