+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.14.2] - Unreleased
+
+### Fixed
+
+- Fix the `Task completed` message during plugin update
+
+### Added
+
+
+### Removed
+
+
 ## [2.14.2] - 2025-08-22
 
 ### Fixed